Implemented move function and tests to check if valid positions result in move
The idea is straightforward, move always moves in the direction the bot
is facing, so just consider that a position to go, check if it's valid
and if so, move.
A clone function had to be added to the kotlin class because the
presenter isn't in kotlin and the much nicer copy function can't be
used.