Right turns were implemented
It was realised that the left and right turns are the same thing just
the turning direction changes so the two functions were dropped and
consolidated into one that does the job of both by just taking in which
direction to turn.
Left side tests were refactored and right side tests were added.